renamed actions/reducers for more consistent naming
