+++ react-ui/src/reducers/index.js
@@ -7,7 +7,7 @@ import registerStudent from './register_student_reducer.js';
 import registerResident from './register_resident_reducer.js';
 import searchStudents from './search_students_reducer.js';
 import applyJob from './apply_job.js';
-import fetchMyJobs from './fetch_my_jobs.js';
+import fetchMyJobs from './my_jobs.js';
 
 export default combineReducers({
     home,
