+++ react-ui/src/reducers/my_jobs.js
